Create hook that parses config to form
